# Conflicts:
#	wwwroot/inc/ajax-interface.php
#	wwwroot/inc/interface-lib.php
#	wwwroot/inc/interface.php
#	wwwroot/inc/navigation.php
#	wwwroot/js/inplace-edit.js

+++ scripts/syncdomain.php
@@ -8,7 +8,7 @@
 $script_mode = TRUE;
 require 'inc/init.php';
 
-function usage()
+function usage ($rc)
 {
 	echo "Usage: <this file> <options>\n";
 	echo "\t\t--vdid=<VLAN domain ID>\n";
@@ -17,17 +17,30 @@ function usage()
 	echo "\t\t--mode=push\n";
 	echo "\t\t[--max=<max_to_do>]\n";
 	echo "\t\t[--verbose]\n";
-	exit (1);
+	echo "\t\t[--nolock]\n";
+	echo "\t\t[--stderr]\n";
+	echo "\t\t--help\n";
+	exit ($rc);
 }
 
-function print_message_line($text)
+define ('PML_VERBOSE', 1 << 0); // display message only if --verbose option specified
+define ('PML_NOTICE',  1 << 1); // the message is informational, do not write to STDERR
+function print_message_line($text, $flags = 0)
 {
-	echo gmdate (DATE_RFC1123) . ": ${text}\n";
+	global $options;
+	if (! array_key_exists ('verbose', $options) and $flags & PML_VERBOSE)
+		return;
+	$buff = date (DATE_RFC1123) . ": ${text}\n";
+	echo $buff;
+	if (array_key_exists ('stderr', $options) and ! ($flags & PML_NOTICE))
+		fwrite (STDERR, $buff);
 }
 
-$options = getopt ('', array ('vdid:', 'max::', 'mode:', 'verbose'));
+$options = getopt ('', array ('vdid:', 'max::', 'mode:', 'verbose', 'nolock', 'stderr', 'help'));
+if (array_key_exists ('help', $options))
+	usage (0);
 if (!array_key_exists ('mode', $options))
-	usage();
+	usage (1);
 
 switch ($options['mode'])
 {
@@ -41,28 +54,11 @@ case 'push':
 	$do_push = TRUE;
 	break;
 default:
-	usage();
+	usage (1);
 }
 
-$max = array_key_exists ('max', $options) ? $options['max'] : 0;
-$verbose = array_key_exists ('verbose', $options);
-
-$switch_list = array();
-if (! isset ($options['vdid']))
-	$switch_list = getVLANSwitches();
-else
-	try
-	{
-		$mydomain = getVLANDomain ($options['vdid']);
-		foreach ($mydomain['switchlist'] as $switch)
-			$switch_list[] = $switch['object_id'];
-	}
-	catch (RackTablesError $e)
-	{
-		print_message_line ("Cannot load domain data with ID ${options['vdid']}");
-		print_message_line ($e->getMessage());
-		exit (1);
-	}
+$max = array_fetch ($options, 'max', 0);
+$nolock = array_key_exists ('nolock', $options);
 
 $todo = array
 (
@@ -71,49 +67,70 @@ $todo = array
 	'pullall' => array ('sync_ready', 'resync_ready', 'sync_aging', 'resync_aging', 'done'),
 );
 
-$domain_key = isset ($options['vdid']) ? $options['vdid'] : 0;
-$filename = '/var/tmp/RackTables-syncdomain-' . $domain_key . '.pid';
-if (FALSE === $fp = @fopen ($filename, 'x+'))
+if (! $nolock)
 {
-	if (FALSE === $pidfile_mtime = filemtime ($filename))
+	$domain_key = isset ($options['vdid']) ? $options['vdid'] : 0;
+	$filename = '/var/tmp/RackTables-syncdomain-' . $domain_key . '.pid';
+	if (FALSE === $fp = @fopen ($filename, 'c+'))
 	{
-		print_message_line ("Failed to obtain mtime of ${filename}");
+		print_message_line ("Failed to open ${filename}");
 		exit (1);
 	}
-	$current_time = time();
-	if ($current_time < $pidfile_mtime)
+	$wouldblock = 0;
+	if (! flock ($fp, LOCK_EX|LOCK_NB, $wouldblock) || $wouldblock)
 	{
-		print_message_line ("Warning: pidfile ${filename} mtime is in future!");
+		$current_time = time();
+		$stat = fstat ($fp);
+		if (! isset ($stat['mtime']))
+		{
+			print_message_line ("Failed to obtain mtime of ${filename}");
+			exit (1);
+		}
+		$pidfile_mtime = $stat['mtime'];
+		if ($current_time < $pidfile_mtime)
+		{
+			print_message_line ("Warning: pidfile ${filename} mtime is in future!");
+			exit (1);
+		}
+		// don't indicate failure unless the pidfile is 15 minutes or more old
+		if ($current_time < $pidfile_mtime + 15 * 60)
+			exit (0);
+		print_message_line ("Failed to lock ${filename}, already locked by PID " . trim (fgets ($fp, 10)));
 		exit (1);
 	}
-	// don't indicate failure unless the pidfile is 15 minutes or more old
-	if ($current_time < $pidfile_mtime + 15 * 60)
-		exit (0);
-	print_message_line ("Failed to lock ${filename}, already locked by PID " . mb_substr (file_get_contents ($filename), 0, 6));
-	exit (1);
+
+	ftruncate ($fp, 0);
+	fwrite ($fp, getmypid() . "\n");
+	// don't close $fp yet: we need to keep an flock
 }
 
-ftruncate ($fp, 0);
-fwrite ($fp, getmypid() . "\n");
-fclose ($fp);
-
 // fetch all the needed data from DB (preparing for DB connection loss)
+$vswitch_filter = array();
+if (isset ($options['vdid']))
+	$vswitch_filter['domain_id'] = $options['vdid'];
+$switch_list = getVLANSwitchInfoRows ($vswitch_filter);
+$enabled_switches = listConstraint ('object', 'SYNC_802Q_LISTSRC');
+
 $switch_queue = array();
-foreach ($switch_list as $object_id)
+foreach ($switch_list as $vswitch)
 {
-	$queue = detectVLANSwitchQueue (getVLANSwitchInfo ($object_id));
-	if ($queue == 'disabled' || in_array ($queue, $todo[$options['mode']]))
+	$object_id = $vswitch['object_id'];
+	$new_disabled = ! isset ($enabled_switches[$object_id]);
+	$queue = detectVLANSwitchQueue ($vswitch);
+	if ($queue != 'disabled' && $new_disabled)
 	{
-		$cell = spotEntity ('object', $object_id);
-		if (considerConfiguredConstraint ($cell, 'SYNC_802Q_LISTSRC'))
-			$switch_queue[] = $cell;
-		else
-			usePreparedExecuteBlade
-			(
-				'UPDATE VLANSwitch SET out_of_sync="yes", last_error_ts=NOW(), last_errno=? WHERE object_id=?',
-				array (E_8021Q_SYNC_DISABLED, $cell['id'])
-			);
+		setVLANSwitchError ($object_id, E_8021Q_SYNC_DISABLED);
+		continue;
 	}
+	elseif ($queue == 'disabled' && ! $new_disabled)
+	{
+		$vswitch['last_errno'] = E_8021Q_NOERROR;
+		setVLANSwitchError ($object_id, $vswitch['last_errno']);
+		$queue = detectVLANSwitchQueue ($vswitch);
+	}
+
+	if (in_array ($queue, $todo[$options['mode']]))
+		$switch_queue[] = spotEntity ('object', $object_id);
 }
 
 // YOU SHOULD NOT USE DB FUNCTIONS BELOW IN THE PARENT PROCESS
@@ -136,7 +153,7 @@ foreach ($switch_queue as $object)
 		}
 		$i_am_child = (0 === $fork_res = pcntl_fork());
 	}
-	if (! $do_fork or $i_am_child)
+	if (! $do_fork || $i_am_child)
 	{
 		try
 		{
@@ -144,8 +161,10 @@ foreach ($switch_queue as $object)
 			if ($i_am_child)
 				connectDB();
 			$portsdone = exec8021QDeploy ($object['id'], $do_push);
-			if ($portsdone or $verbose)
-				print_message_line ("Done '${object['dname']}': ${portsdone}");
+			$flags = PML_NOTICE;
+			if (! $portsdone)
+				$flags |= PML_VERBOSE;
+			print_message_line ("Done '${object['dname']}': ${portsdone}", $flags);
 		}
 		catch (RackTablesError $e)
 		{
@@ -159,8 +178,7 @@ foreach ($switch_queue as $object)
 
 	if (++$switchesdone == $max)
 	{
-		if ($verbose)
-			print_message_line ("Maximum of ${max} items reached, terminating");
+		print_message_line ("Maximum of ${max} items reached, terminating", PML_NOTICE|PML_VERBOSE);
 		break;
 	}
 }
@@ -172,10 +190,14 @@ while ($switches_working > 0)
 	pcntl_waitpid (-1, $wait_status);
 }
 
-if (FALSE === unlink ($filename))
+if (! $nolock)
 {
-	print_message_line ("Failed removing pidfile ${filename}");
-	exit (1);
+	flock ($fp, LOCK_UN); // explicitly unlock file as PHP 5.3.2 made it mandatory
+	if (FALSE === unlink ($filename))
+	{
+		print_message_line ("Failed removing pidfile ${filename}");
+		exit (1);
+	}
 }
 exit (0);
 ?>
